ADDED: Added a Description tag in the `Dgmjr.Primitives.All.csproj` file to provide a description for the metapackage.
UPDATED: Updated the `README.md` file in the `All` directory to include a description for the metapackage.

UPDATED: Updated the `DayTimeDuration.cs` file to change the type of `TicksPerSecond` and `TicksPerMillisecond` constants to `decimal`.

UPDATED: Updated the `uri.cs` file to include a description for the uri type.

ADDED: Added the `FluentDateTime` package reference in the `Dgmjr.Primitives.csproj` file.

UPDATED: Updated the `url.cs` file to include a description for the url type.

UPDATED: Updated the `urn.cs` file to include a description for the urn type.

UPDATED: Updated the `xri.cs` file to include a description for the xri type.